// frontend/src/types/data-prm.types/category-select.type.d.ts
export declare enum CategoryTypes {
    cost = "expense",
    income = "income"
}
export declare class CategorySwodRecord {
    type: CategoryTypes;
    date: string | null;
    category_id: number | null;
    amount: number | null;
    comment: string | null;
    constructor();
}
//# sourceMappingURL=category-select.type.d.ts.map